fix: crash on 'version' command at opensuse platform
This issue may not be only on opensuse platform, ever platform has
potential to turn out this bug. The root cause is a sloopy pointer
handling for parsing the cpu string. On systems where one item of
/proc/cpuinfo goes over one line (meaning that the searched ':' doesn't
occur in the string) "pk" is left uninitalized, so the access of strchr
crashes.

Signed-off-by: Hannes Schmelzer <oe5hpm@oevsv.at>